[refactor] Vote 패키지 구조 및 테이블 이름 변경 (#80)
