+++ src/test/java/konkuk/thip/record/adapter/in/web/RecordQueryControllerTest.java
@@ -19,8 +19,8 @@ import konkuk.thip.user.adapter.out.persistence.repository.AliasJpaRepository;
 import konkuk.thip.user.adapter.out.persistence.repository.UserJpaRepository;
 import konkuk.thip.vote.adapter.out.jpa.VoteItemJpaEntity;
 import konkuk.thip.vote.adapter.out.jpa.VoteJpaEntity;
-import konkuk.thip.vote.adapter.out.persistence.VoteItemJpaRepository;
-import konkuk.thip.vote.adapter.out.persistence.VoteJpaRepository;
+import konkuk.thip.vote.adapter.out.persistence.repository.VoteItemJpaRepository;
+import konkuk.thip.vote.adapter.out.persistence.repository.VoteJpaRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
