+++ src/test/java/konkuk/thip/room/adapter/in/web/RoomGetMemberListApiTest.java
@@ -11,7 +11,7 @@ import konkuk.thip.user.adapter.out.jpa.AliasJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserJpaEntity;
 import konkuk.thip.room.adapter.out.jpa.RoomParticipantRole;
 import konkuk.thip.user.adapter.out.persistence.repository.AliasJpaRepository;
-import konkuk.thip.user.adapter.out.persistence.repository.FollowingJpaRepository;
+import konkuk.thip.user.adapter.out.persistence.repository.following.FollowingJpaRepository;
 import konkuk.thip.user.adapter.out.persistence.repository.UserJpaRepository;
 import konkuk.thip.room.adapter.out.persistence.repository.RoomParticipantJpaRepository;
 import org.junit.jupiter.api.AfterEach;
